Improved visuals + small bug fix
Previously, the whole list was printing when being called, e.g: "("Nuggets", "Butter", "Milk"...)". I improved the look so that it instead prints, "Nuggets. Butter. Milk." (without the speech marks to be clear). Also, there was a small bug where the program would ask if the user would like to add more ingredients to the initial recipe question, and if they responded improperly to a yes or no question it would just fall through and skip instead of repeating the question until getting a yes/no response.